requirements.txt
Flask==2.3.2
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Jinja2==3.1.2
mysql-connector-python==8.0.33
pymysql==1.0.3
Werkzeug==2.3.6
python-dotenv==1.0.0
WTForms==3.0.1
